Add a tap function
This function lazily performs a side effect for each item in an iterable,
without changing the keys/values of said iterable.

This is useful for things like logging, saving partial results to a database, or
any other side effects which should not change the outcome of the full iteration
pipeline.